+++ src/main/java/com/itlize/res/repository/ProjectColumnsRepository.java
@@ -4,10 +4,10 @@ import com.itlize.res.entity.Project;
 import com.itlize.res.entity.ProjectColumns;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectColumnsRepository extends JpaRepository<ProjectColumns, Integer> {
+public interface ProjectColumnsRepository extends JpaRepository<ProjectColumns, Long> {
 
-    ProjectColumns findProjectColumnsByColumnId(Integer columnId);
-    String deleteProjectColumnsByColumnId(Integer id);
-    ProjectColumns findByColumnId(Integer columnid);
-    Boolean existsByColumnId(Integer column_id);
+    ProjectColumns findProjectColumnsByColumnId(Long columnId);
+    String deleteProjectColumnsByColumnId(Long id);
+    ProjectColumns findByColumnId(Long columnid);
+    Boolean existsByColumnId(Long column_id);
 }
