+++ src/main/java/com/itlize/res/repository/ProjectRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProjectRepository extends JpaRepository<Project, Integer> {
+public interface ProjectRepository extends JpaRepository<Project, Long> {
     //void deleteProjectRepositoryById(Integer id);
 
-    Boolean existsByProjectId(Integer id);
-    String deleteByProjectId(Integer id);
-    Project findByProjectId(Integer id);
+    Boolean existsByProjectId(Long id);
+    String deleteByProjectId(Long id);
+    Project findByProjectId(Long id);
 }
