+++ src/main/java/com/itlize/res/repository/ProjectResourceRepository.java
@@ -4,7 +4,7 @@ import com.itlize.res.entity.ProjectResource;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface ProjectResourceRepository extends JpaRepository<ProjectResource, Integer> {
+public interface ProjectResourceRepository extends JpaRepository<ProjectResource, Long> {
 
 //    void findProjectResourceByResource();
 //
