+++ src/main/java/com/itlize/res/repository/ResourceDetailsRepository.java
@@ -4,7 +4,7 @@ import com.itlize.res.entity.Project;
 import com.itlize.res.entity.ResourceDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResourceDetailsRepository extends JpaRepository<ResourceDetails, Integer> {
-    Boolean existsByRecordId(Integer id);
-    ResourceDetails findByRecordId(Integer id);
+public interface ResourceDetailsRepository extends JpaRepository<ResourceDetails, Long> {
+    Boolean existsByRecordId(Long id);
+    ResourceDetails findByRecordId(Long id);
 }
