+++ src/main/java/com/itlize/res/repository/UserRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findAllByUsernameOrEmail(String userName, String email);
  
@@ -17,7 +17,7 @@ public interface UserRepository extends JpaRepository<User, Integer> {
 
 
     //List<User> findByUserIDIn(List<Integer> userIds);
-    User findByUserId(Integer id);
+    User findByUserId(Long id);
